added working like & comment buttons

+++ app/routes.py
@@ -2,7 +2,7 @@ from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import login_user, login_required, logout_user, current_user
 from werkzeug.security import generate_password_hash, check_password_hash
 from app import db
-from app.models import User, Blog
+from app.models import User, Blog, Comment  # Import Comment model
 
 main = Blueprint('main', __name__)
 
@@ -40,8 +40,6 @@ def login():
 
     return render_template('login.html')
 
-
-
 # Home route
 @main.route('/')
 def home():
@@ -64,34 +62,57 @@ def create_blog():
 
     return render_template('create_blog.html')
 
-# View blog route
-@main.route("/blog/<int:blog_id>")
+# View blog and handle comments
+@main.route("/blog/<int:blog_id>", methods=['GET', 'POST'])
 def view_blog(blog_id):
     blog = Blog.query.get_or_404(blog_id)
+    
+    # Handle comment form submission
+    if request.method == 'POST':
+        if current_user.is_authenticated:
+            comment_content = request.form.get('content')
+            if comment_content:
+                comment = Comment(content=comment_content, user_id=current_user.id, blog_id=blog.id)
+                db.session.add(comment)
+                db.session.commit()
+                flash('Comment added!', 'success')
+            else:
+                flash('Comment cannot be empty.', 'warning')
+        else:
+            flash('You must be logged in to comment.', 'danger')
+        return redirect(url_for('main.view_blog', blog_id=blog.id))
+    
     return render_template('view_blog.html', blog=blog)
 
 # Edit blog route
 @main.route('/blog/edit/<int:blog_id>', methods=['GET', 'POST'])
+@login_required
 def edit_blog(blog_id):
     blog = Blog.query.get_or_404(blog_id)
     
     if request.method == 'POST':
-        blog.title = request.form['title']
-        blog.content = request.form['content']
-        
-        db.session.commit()
-        flash('Blog Updated Successfully!', 'success')
-        return redirect(url_for('main.home'))
+        if blog.user_id == current_user.id or current_user.is_admin:
+            blog.title = request.form['title']
+            blog.content = request.form['content']
+            db.session.commit()
+            flash('Blog Updated Successfully!', 'success')
+            return redirect(url_for('main.home'))
+        else:
+            flash('You are not authorized to edit this blog.', 'danger')
 
     return render_template('edit_blog.html', blog=blog)
 
 # Delete blog route
 @main.route('/blog/delete/<int:blog_id>', methods=['POST'])
+@login_required
 def delete_blog(blog_id):
     blog = Blog.query.get_or_404(blog_id)
-    db.session.delete(blog)
-    db.session.commit()
-    flash('Blog Deleted Successfully!', 'danger')
+    if blog.user_id == current_user.id or current_user.is_admin:
+        db.session.delete(blog)
+        db.session.commit()
+        flash('Blog Deleted Successfully!', 'danger')
+    else:
+        flash('You are not authorized to delete this blog.', 'danger')
     return redirect(url_for('main.home'))
 
 # Admin panel route
@@ -117,7 +138,6 @@ def delete_blog_from_admin(blog_id):
     else:
         flash("You are not authorized to delete this blog.", "danger")
     return redirect(url_for('main.admin'))
-from flask_login import logout_user
 
 # Logout route
 @main.route('/logout')
@@ -126,3 +146,13 @@ def logout():
     logout_user()
     flash("You have been logged out.", "info")
     return redirect(url_for('main.login'))
+
+# Like a blog
+@main.route('/like/<int:blog_id>', methods=['POST'])
+@login_required
+def like_blog(blog_id):
+    blog = Blog.query.get_or_404(blog_id)
+    blog.likes += 1
+    db.session.commit()
+    flash('You liked this blog!', 'success')
+    return redirect(url_for('main.view_blog', blog_id=blog.id))
